requirements.txt
flask
mysql-connector-python
qrcode[pil]
werkzeug
gunicorn
